+++ packages/terra-date-picker/src/react-datepicker/index.jsx
@@ -1,6 +1,9 @@
 import Calendar from './calendar'
 import React from 'react'
 import PropTypes from 'prop-types'
+import FocusTrap from 'focus-trap-react';
+import { Portal } from 'react-portal';
+import KeyCode from 'keycode-js';
 import Hookshot from 'terra-hookshot'
 import classnames from 'classnames'
 import {
@@ -149,8 +152,27 @@ export default class DatePicker extends React.Component {
     super(props)
     this.state = this.calcInitialState()
     this.handleOnPosition = this.handleOnPosition.bind(this);
+    this.handleKeydown = this.handleKeydown.bind(this);
     this.datePickerContainer = React.createRef();
     this.datePickerHookShotContainer = React.createRef();
+    this.datePickerOverlayContainer = React.createRef();
+    this.handleCalendarKeyDown = this.handleCalendarKeyDown.bind(this);
+  }
+
+  componentDidMount() {
+    document.addEventListener('keydown', this.handleKeydown);
+  }
+
+  componentDidUpdate() {
+    // Shift focus into hookshot date-picker if it exists
+    if (this.datePickerHookShotContainer.current) {
+      this.datePickerHookShotContainer.current.focus();
+    }
+
+    // Shift focus into overlay date-picker if it exists
+    if (this.datePickerOverlayContainer.current) {
+      this.datePickerOverlayContainer.current.focus();
+    }
   }
 
   componentWillReceiveProps (nextProps) {
@@ -165,7 +187,26 @@ export default class DatePicker extends React.Component {
   }
 
   componentWillUnmount () {
+    document.removeEventListener('keydown', this.handleKeydown);
     this.clearPreventFocusTimeout()
+  }
+
+  handleKeydown(event) {
+    if (event.keyCode === KeyCode.KEY_ESCAPE) {
+      // If date picker is open in Hookshot
+      if (this.datePickerHookShotContainer.current) {
+        if (event.target === this.datePickerHookShotContainer.current || this.datePickerHookShotContainer.current.contains(event.target)) {
+          this.setOpen(false);
+        }
+      }
+
+      // If date picker is open in overlay
+      if (this.datePickerOverlayContainer.current) {
+        if (event.target === this.datePickerOverlayContainer.current || this.datePickerOverlayContainer.current.contains(event.target)) {
+          this.setOpen(false);
+        }
+      }
+    }
   }
 
   getPreSelection = () => (
@@ -359,15 +400,9 @@ export default class DatePicker extends React.Component {
     }
   }
 
-  onInputKeyDown = (event) => {
+  handleCalendarKeyDown = (event) => {
     this.props.onKeyDown(event)
     const eventKey = event.key
-    if (!this.state.open && !this.props.inline) {
-      if (eventKey !== 'Enter' && eventKey !== 'Escape' && eventKey !== 'Tab') {
-        this.onInputClick()
-      }
-      return
-    }
     const copy = newDate(this.state.preSelection)
     if (eventKey === 'Enter') {
       event.preventDefault()
@@ -552,9 +587,18 @@ export default class DatePicker extends React.Component {
           }
           {
             this.state.open || this.props.inline
-              ? <div className="react-datepicker__portal">
-                {calendar}
-              </div>
+              ? (<Portal isOpened={true}>
+                  <FocusTrap focusTrapOptions={{ returnFocusOnDeactivate: true, clickOutsideDeactivates: true }}>
+                    <div
+                      ref={this.datePickerOverlayContainer}
+                      className="react-datepicker__portal"
+                      tabIndex="-1"
+                      onKeyDown={this.handleCalendarKeyDown}
+                    >
+                      {calendar}
+                    </div>
+                  </FocusTrap>
+                </Portal>)
               : null
           }
         </div>
@@ -580,13 +624,19 @@ export default class DatePicker extends React.Component {
           onPosition={this.handleOnPosition}
         >
           <Hookshot.Content>
-            <div
-              className="react-datepicker-hookshot"
-              data-placement="bottom"
-              ref={this.datePickerHookShotContainer}
-            >
-              {calendar}
-            </div>
+            <FocusTrap focusTrapOptions={{ returnFocusOnDeactivate: true, clickOutsideDeactivates: true }}>
+              <div>
+                <div
+                  className="react-datepicker-hookshot"
+                  data-placement="bottom"
+                  ref={this.datePickerHookShotContainer}
+                  tabIndex="-1"
+                  onKeyDown={this.handleCalendarKeyDown}
+                >
+                  {calendar}
+                </div>
+              </div>
+            </FocusTrap>
           </Hookshot.Content>
         </Hookshot>}
       </React.Fragment>