+++ packages/terra-notification-dialog/CHANGELOG.md
@@ -3,6 +3,13 @@ Changelog
 
 Unreleased
 ----------
+### Breaking Changes
+* Removed `releaseFocus` prop
+* Removed `requestFocus` prop
+
+### Added
+* Added component specific handler for Escape key usage with Notification Dialog
+* Added focus trap to Notification Dialog
 
 2.12.0 - (May 7, 2019)
 ------------------